Multi client

+++ Classes/Controller/PositionController.php
@@ -33,7 +33,7 @@ namespace Webfox\Placements\Controller;
  * @license http://www.gnu.org/licenses/gpl.html GNU General Public License, version 3 or later
  *
  */
-class PositionController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionController {
+class PositionController extends AbstractController {
 
 	/**
 	 * Position Repository
@@ -52,36 +52,12 @@ class PositionController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControl
 	protected $positionTypeRepository;
 
 	/**
-	 * Sector Repository
-	 *
-	 * @var \Webfox\Placements\Domain\Repository\SectorRepository
-	 * @inject
-	 */
-	protected $sectorRepository;
-
-	/**
-	 * Organization Repository
-	 *
-	 * @var \Webfox\Placements\Domain\Repository\OrganizationRepository
-	 * @inject
-	 */
-	protected $organizationRepository;
-
-	/**
 	 * Working Hours Repository
 	 *
 	 * @var \Webfox\Placements\Domain\Repository\WorkingHoursRepository
 	 * @inject
 	 */
 	protected $workingHoursRepository;
-
-	/**
-	 * Category Repository
-	 *
-	 * @var \Webfox\Placements\Domain\Repository\CategoryRepository
-	 * @inject
-	 */
-	protected $categoryRepository;
 
 	/**
 	 * Initialize Action
@@ -118,6 +94,7 @@ class PositionController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControl
 			);
 		}
 	}
+
 	/**
 	 * action list
 	 *
@@ -151,19 +128,33 @@ class PositionController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControl
 	 * @return void
 	 */
 	public function newAction(\Webfox\Placements\Domain\Model\Position $newPosition = NULL) {
-		$positionTypes = $this->positionTypeRepository->findMultipleByUid($this->settings['positionTypes'], 'title');
-		$workingHours = $this->workingHoursRepository->findMultipleByUid($this->settings['workingHours'], 'title');
-		$categories = $this->categoryRepository->findMultipleByUid($this->settings['categories'], 'title');
-		$sectors = $this->sectorRepository->findMultipleByUid($this->settings['sectors'], 'title');
-		$organizations = $this->organizationRepository->findAll();
-		$this->view->assignMultiple(array(
-			'newPosition' => $newPosition,
-			'workingHours' => $workingHours,
-			'positionTypes' => $positionTypes,
-			'categories' => $categories,
-			'sectors' => $sectors,
-			'organizations' => $organizations,
-		));
+		if(!$this->accessControlService->isAllowedToCreate('position')) {
+			$this->flashMessageContainer->add(
+				\TYPO3\CMS\Extbase\Utility\LocalizationUtility::translate(
+					'tx_placements.error.position.createActionNotAllowed', 'placements'
+				)
+			);
+			$this->redirect('list');
+		} else {
+			$positionTypes = $this->positionTypeRepository->findMultipleByUid($this->settings['positionTypes'], 'title');
+			$workingHours = $this->workingHoursRepository->findMultipleByUid($this->settings['workingHours'], 'title');
+			$categories = $this->categoryRepository->findMultipleByUid($this->settings['categories'], 'title');
+			$sectors = $this->sectorRepository->findMultipleByUid($this->settings['sectors'], 'title');
+			$user = $this->accessControlService->getFrontendUser();
+			if ($user AND $user->getClient()) {
+			    $organizations = $this->organizationRepository->findByClient($user->getClient());
+			} else {
+				$organizations = $this->organizationRepository->findAll();
+			}
+			$this->view->assignMultiple(array(
+				'newPosition' => $newPosition,
+				'workingHours' => $workingHours,
+				'positionTypes' => $positionTypes,
+				'categories' => $categories,
+				'sectors' => $sectors,
+				'organizations' => $organizations,
+			));
+		}
 	}
 
 	/**
@@ -173,7 +164,8 @@ class PositionController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControl
 	 * @return void
 	 */
 	public function createAction(\Webfox\Placements\Domain\Model\Position $newPosition) {
-		$this->positionRepository->add($newPosition);
+		$newPosition->setClient($this->accessControlService->getFrontendUser()->getClient());
+	    	$this->positionRepository->add($newPosition);
 		$this->flashMessageContainer->add(
 			\TYPO3\CMS\Extbase\Utility\LocalizationUtility::translate(
 				'tx_placements.success.position.createAction', 'placements'
@@ -190,19 +182,32 @@ class PositionController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControl
 	 * @return void
 	 */
 	public function editAction(\Webfox\Placements\Domain\Model\Position $position) {
-		$positionTypes = $this->positionTypeRepository->findMultipleByUid($this->settings['positionTypes'], 'title');
-		$workingHours = $this->workingHoursRepository->findMultipleByUid($this->settings['workingHours'], 'title');
-		$categories = $this->categoryRepository->findMultipleByUid($this->settings['categories'], 'title');
-		$sectors = $this->sectorRepository->findMultipleByUid($this->settings['sectors'], 'title');
-		$organizations = $this->organizationRepository->findAll();
-		$this->view->assignMultiple(array(
-			'position' => $position,
-			'workingHours' => $workingHours,
-			'positionTypes' => $positionTypes,
-			'categories' => $categories,
-			'sectors' => $sectors,
-			'organizations' => $organizations,
-		));
+		if(!$this->accessControlService->isAllowedToEdit('position')) {
+			$this->flashMessageContainer->add(
+				\TYPO3\CMS\Extbase\Utility\LocalizationUtility::translate(
+					'tx_placements.error.position.editActionNotAllowed', 'placements'
+				)
+			);
+			$this->redirect('list');
+		} else {
+			$positionTypes = $this->positionTypeRepository->findMultipleByUid($this->settings['positionTypes'], 'title');
+			$workingHours = $this->workingHoursRepository->findMultipleByUid($this->settings['workingHours'], 'title');
+			$categories = $this->categoryRepository->findMultipleByUid($this->settings['categories'], 'title');
+			$sectors = $this->sectorRepository->findMultipleByUid($this->settings['sectors'], 'title');
+			if ($user AND $user->getClient()) {
+			    $organizations = $this->organizationRepository->findByClient($user->getClient());
+			} else {
+				$organizations = $this->organizationRepository->findAll();
+			}
+			$this->view->assignMultiple(array(
+				'position' => $position,
+				'workingHours' => $workingHours,
+				'positionTypes' => $positionTypes,
+				'categories' => $categories,
+				'sectors' => $sectors,
+				'organizations' => $organizations,
+			));
+		}
 	}
 
 	/**
@@ -212,9 +217,13 @@ class PositionController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControl
 	 * @return void
 	 */
 	public function updateAction(\Webfox\Placements\Domain\Model\Position $position) {
-		$this->updateStorageProperties($position);
+		//$this->updateStorageProperties($position);
 		$this->positionRepository->update($position);
-		$this->flashMessageContainer->add('Your Position was updated.');
+		$this->flashMessageContainer->add(
+			\TYPO3\CMS\Extbase\Utility\LocalizationUtility::translate(
+				'tx_placements.success.position.updateAction', 'placements'
+				)
+		);
 		$this->redirect('list');
 	}
 
@@ -225,8 +234,20 @@ class PositionController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControl
 	 * @return void
 	 */
 	public function deleteAction(\Webfox\Placements\Domain\Model\Position $position) {
-		$this->positionRepository->remove($position);
-		$this->flashMessageContainer->add('Your Position was removed.');
+		if($this->accessControlService->isAllowedToDelete('position')) {
+			$this->positionRepository->remove($position);
+			$this->flashMessageContainer->add(
+				\TYPO3\CMS\Extbase\Utility\LocalizationUtility::translate(
+					'tx_placements.success.position.deleteAction', 'placements'
+				)
+			);
+		} else {
+			$this->flashMessageContainer->add(
+				\TYPO3\CMS\Extbase\Utility\LocalizationUtility::translate(
+					'tx_placements.error.position.deleteActionNotAllowed', 'placements'
+				)
+			);
+		}
 		$this->redirect('list');
 	}
 
@@ -353,6 +374,13 @@ class PositionController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControl
 		if($settings['categoryConjunction'] !== '') {
 			$demand->setCategoryConjunction($settings['categoryConjunction']);
 		}
+		if($settings['clientsPositionsOnly'] AND 
+			$this->accessControlService->hasLoggedInClient()) {
+			$demand->setClients(
+							$this->accessControlService->
+							getFrontendUser()->getClient()->getUid());
+			$demand->setClientsPositionsOnly(TRUE);
+		}
 		$demand->setLimit($settings['limit']);
 		return $demand;
 	}
@@ -366,12 +394,25 @@ class PositionController extends \TYPO3\CMS\Extbase\Mvc\Controller\ActionControl
 	 */
 	protected function overwriteDemandObject($demand, $overwriteDemand) {
 		unset($overwriteDemand['orderByAllowed']);
+		if($overwriteDemand['clientsPositionsOnly'] AND
+			$this->accessControlService->hasLoggedInClient()) {
+			$demand->setClients(
+							(string)
+							$this->
+							accessControlService->
+							getFrontendUser()->
+							getClient()->getUid());
+		} elseif ($overwriteDemand['clientsPositionOnly'] == '') {
+			$demand->setClients('');
+		}
+			
 
 		foreach ($overwriteDemand as $propertyName => $propertyValue) {
 			if(!empty($propertyValue)) {
 				\TYPO3\CMS\Extbase\Reflection\ObjectAccess::setProperty($demand, $propertyName, $propertyValue);
 		    }
 		}
+
 		//store session data
 		/*
 		$sessionData = serialize($overwriteDemand);
